src/requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.in
#
aiosqlite==0.16.1
    # via tortoise-orm
appdirs==1.4.4
    # via black
asgiref==3.4.1
    # via uvicorn
asyncpg==0.23.0
    # via -r requirements/base.in
attrs==21.2.0
    # via pytest
black==21.7b0
    # via -r requirements/dev.in
certifi==2021.5.30
    # via requests
charset-normalizer==2.0.3
    # via requests
click==8.0.1
    # via
    #   black
    #   uvicorn
coverage==5.5
    # via pytest-cov
fastapi-camelcase==1.0.2
    # via -r requirements/base.in
fastapi-utils==0.2.1
    # via -r requirements/base.in
fastapi==0.66.1
    # via
    #   -r requirements/base.in
    #   fastapi-utils
flake8==3.9.2
    # via -r requirements/dev.in
greenlet==1.1.0
    # via sqlalchemy
h11==0.12.0
    # via uvicorn
idna==3.2
    # via requests
iniconfig==1.1.1
    # via pytest
iso8601==0.1.16
    # via tortoise-orm
isort==5.9.2
    # via -r requirements/dev.in
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
packaging==21.0
    # via pytest
pathspec==0.9.0
    # via black
pluggy==0.13.1
    # via pytest
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pydantic==1.8.2
    # via
    #   fastapi
    #   fastapi-camelcase
    #   fastapi-utils
pyflakes==2.3.1
    # via flake8
pyhumps==3.0.2
    # via fastapi-camelcase
pyparsing==2.4.7
    # via packaging
pypika-tortoise==0.1.1
    # via tortoise-orm
pytest-asyncio==0.15.1
    # via -r requirements/dev.in
pytest-cov==2.12.1
    # via -r requirements/dev.in
pytest==6.2.4
    # via
    #   -r requirements/dev.in
    #   pytest-asyncio
    #   pytest-cov
python-multipart==0.0.5
    # via -r requirements/base.in
pytz==2021.1
    # via tortoise-orm
regex==2021.7.6
    # via black
requests==2.26.0
    # via -r requirements/dev.in
six==1.16.0
    # via python-multipart
sqlalchemy==1.4.21
    # via fastapi-utils
starlette==0.14.2
    # via fastapi
toml==0.10.2
    # via
    #   pytest
    #   pytest-cov
tomli==1.0.4
    # via black
tortoise-orm==0.17.5
    # via -r requirements/base.in
typing-extensions==3.10.0.0
    # via
    #   aiosqlite
    #   pydantic
urllib3==1.26.6
    # via requests
uvicorn==0.14.0
    # via -r requirements/base.in
